feat: 회원탈퇴 플로우 확인
