+++ DonDog-iOS/DonDog-iOS/App/Factory/ModuleFactoryProtocol.swift
@@ -11,7 +11,7 @@ import Combine
 protocol ModuleFactoryProtocol {
     func makeWelcomeView() -> WelcomeView
     func makeAuthView(isWithDraw: Bool) -> AuthView
-    func makeAuthNumberView() -> AuthNumberView
+    func makeAuthNumberView(isNumberWithdraw: Bool) -> AuthNumberView
     func makeProfileSetupView() -> ProfileSetupView
     func makeInviteView(showSentHint: Bool) -> InviteView
     func makeCameraView(with feedViewModel: FeedViewModel) -> CameraView
@@ -37,8 +37,8 @@ final class ModuleFactory: ModuleFactoryProtocol {
         return view
     }
     
-    func makeAuthNumberView() -> AuthNumberView {
-        let viewModel = AuthNumberViewModel()
+    func makeAuthNumberView(isNumberWithdraw: Bool) -> AuthNumberView {
+        let viewModel = AuthNumberViewModel(isNumberWithdraw: isNumberWithdraw)
         let view = AuthNumberView(viewModel: viewModel)
         return view
     }
