+++ DonDog-iOS/DonDog-iOS/Presentation/Features/Auth/Views/AuthNumberView.swift
@@ -14,7 +14,9 @@ struct AuthNumberView: View {
     
     var body: some View {
         VStack(spacing: 0) {
-            CustomNavigationBar(leadingType: .back(action: { coordinator.pop() }), centerType: .title(title: "본인인증"), trailingType: .none, navigationColor: .black)
+            CustomNavigationBar(leadingType: .back(action: {
+                coordinator.pop()
+            }), centerType: .title(title: "본인인증"), trailingType: .none, navigationColor: .black)
             
             Spacer()
                 .frame(height: 104)
@@ -55,10 +57,5 @@ struct AuthNumberView: View {
         .padding(.horizontal, 20)
         .backHiddenSwipeEnabled()
         .dismissKeyboard()
-        
     }
-}
-
-#Preview {
-    AuthNumberView(viewModel: AuthNumberViewModel())
 }
